fix: No such file or directory

+++ scripts/bypass-another-branch.sh
@@ -36,6 +36,10 @@ git -C "$LOCAL_REPO" branch --remotes | grep -vF 'HEAD ->' | cut -d ' ' -f 3 | w
 
   git -C "$LOCAL_REPO" checkout --quiet "$branch"
 
+  if ! test -f "$LOCAL_REPO/.github/workflows"; then
+    continue
+  done
+
   find "$LOCAL_REPO/.github/workflows" -type f | grep -E '\.ya?ml$' | while read file; do
     cp "$file" "$OUTPUT_DIR/$branch/.github/workflows"
   done
